Simplified base assemblers by removing unnecessary code
