+++ specs/pom.xml
@@ -40,6 +40,11 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>org.easytesting</groupId>
+            <artifactId>fest-reflect</artifactId>
+            <version>1.4</version>
+        </dependency>
 
         <dependency>
             <groupId>org.seedstack.seed</groupId>
