+++ specs/src/main/java/org/seedstack/business/api/Tuples.java
@@ -7,7 +7,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.business.helpers;
+package org.seedstack.business.api;
 
 import com.google.common.collect.Lists;
 import com.google.inject.util.Types;
